requirements.txt
django
djangorestframework
packaging